backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Data & spatial
pandas==2.2.3
geopandas==1.0.1
shapely==2.0.6
pyproj==3.7.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
supabase==2.11.0

# Google Earth Engine
earthengine-api==1.4.2

# HTTP & async
httpx==0.28.1
aiohttp==3.11.10

# Data validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
redis==5.2.1
celery==5.4.0

# AI
google-generativeai==0.8.3

# News aggregation
feedparser==6.0.11

# Phase 2: file upload handling
python-multipart==0.0.20

# Phase 2: spatial accuracy (Bhuvan WMS / OWS)
OWSLib==0.31.0
